refactor: modularize agent scripts
- Split agent-readme.ts into scripts/agent/ modules:
  - utils.ts: shared constants and helpers
  - roadmap.ts: GitHub Project/Issue management
  - epic.ts: Epic issue creation and tracking
  - whitebook.ts: documentation navigation
  - readme.ts: index output functions
  - cli.ts: main CLI entry point

- Add Epic management commands:
  - pnpm agent --epic create <title>
  - pnpm agent --epic list
  - pnpm agent --epic view <#>
  - pnpm agent --epic sync <#>
  - pnpm agent --epic add <epic#> <issue#>

- Update CLAUDE.md to use pnpm agent
